Move tracker out of `__init__.py`
